requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo>=4.9,<4.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
email-validator==2.2.0
jinja2==3.1.4
python-dotenv==1.0.1
gunicorn==23.0.0
